Move WinDialog to interfaces sub-module.
